Move tests closer to real SUT
