+++ tests/SimpleAccounting.UnitTests/Model/ProjectDataTests.cs
@@ -7,10 +7,12 @@ namespace lg2de.SimpleAccounting.UnitTests.Model;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 using Caliburn.Micro;
 using FluentAssertions;
+using FluentAssertions.Execution;
 using FluentAssertions.Extensions;
 using lg2de.SimpleAccounting.Abstractions;
 using lg2de.SimpleAccounting.Infrastructure;
@@ -23,6 +25,219 @@ using Xunit;
 
 public class ProjectDataTests
 {
+    [Fact]
+    public async Task LoadFromFileAsync_HappyPath_FileLoaded()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        fileSystem.FileExists("the.fileName").Returns(true);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        using var _ = new AssertionScope();
+        sut.FileName.Should().Be("the.fileName");
+        sut.IsModified.Should().BeFalse();
+        sut.Settings.RecentProject.Should().Be("the.fileName");
+        sut.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
+        fileSystem.Received(1).ReadAllTextFromFile("the.fileName");
+    }
+
+    [CulturedFact("en")]
+    public async Task LoadFromFileAsync_UserWantsAutoSaveFile_AutoSaveFileLoaded()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        dialogs.ShowMessageBox(
+                Arg.Is<string>(
+                    s => s.Contains("automatically generated backup file for project", StringComparison.Ordinal)),
+                Arg.Any<string>(),
+                MessageBoxButton.YesNo, MessageBoxImage.Question,
+                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
+            .Returns(MessageBoxResult.Yes);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+        fileSystem.FileExists("the.fileName").Returns(true);
+        fileSystem.FileExists("the.fileName~").Returns(true);
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(5.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        using var _ = new AssertionScope();
+        sut.FileName.Should().Be("the.fileName");
+        sut.IsModified.Should().BeTrue("changes are (still) not yet saved");
+        sut.Settings.RecentProject.Should().Be("the.fileName");
+        sut.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
+        fileSystem.Received(1).ReadAllTextFromFile("the.fileName~");
+    }
+
+    [CulturedFact("en")]
+    public async Task LoadFromFileAsync_UserDoesNotWantAutoSaveFileExists_ProjectFileLoaded()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        dialogs.ShowMessageBox(
+                Arg.Is<string>(
+                    s => s.Contains("automatically generated backup file for project", StringComparison.Ordinal)),
+                Arg.Any<string>(),
+                MessageBoxButton.YesNo, MessageBoxImage.Question,
+                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
+            .Returns(MessageBoxResult.No);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+        fileSystem.FileExists("the.fileName").Returns(true);
+        fileSystem.FileExists("the.fileName~").Returns(true);
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        using var _ = new AssertionScope();
+        sut.FileName.Should().Be("the.fileName");
+        sut.IsModified.Should().BeFalse();
+        sut.Settings.RecentProject.Should().Be("the.fileName");
+        sut.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
+        fileSystem.Received(1).ReadAllTextFromFile("the.fileName");
+        fileSystem.Received(1).FileDelete("the.fileName~");
+    }
+
+    [Theory]
+    [InlineData("Cryptomator File System")]
+    [InlineData("cryptoFs")]
+    public async Task LoadProjectFromFileAsync_NewFileOnSecureDrive_StoreOpenedAndFileLoaded(
+        string cryptoDriveIdentifier)
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        fileSystem.FileExists(Arg.Is("K:\\the.fileName")).Returns(true);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+        fileSystem.GetDrives().Returns(
+            _ =>
+            {
+                var func1 = new Func<string>(() => "Normal");
+                var func2 = new Func<string>(() => cryptoDriveIdentifier);
+                return new[] { (FilePath: "C:\\", GetFormat: func1), (FilePath: "K:\\", GetFormat: func2) };
+            });
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("K:\\the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        sut.Settings.SecuredDrives.Should().BeEquivalentTo(new object[] { "K:\\" });
+        dialogs.DidNotReceive().ShowMessageBox(
+            Arg.Is<string>(s => s.Contains("Cryptomator", StringComparison.Ordinal)),
+            Arg.Any<string>(),
+            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
+        fileSystem.Received(1).ReadAllTextFromFile("K:\\the.fileName");
+    }
+
+    [Fact]
+    public async Task LoadProjectFromFileAsync_FullRecentList_NewFileOnTop()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi)
+        {
+            Settings =
+            {
+                RecentProjects =
+                [
+                    "A",
+                    "B",
+                    "C",
+                    "D",
+                    "E",
+                    "F",
+                    "G",
+                    "H",
+                    "I",
+                    "J"
+                ]
+            }
+        };
+        fileSystem.FileExists("the.fileName").Returns(true);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        sut.Settings.RecentProjects.OfType<string>().Should()
+            .Equal("the.fileName", "A", "B", "C", "D", "E", "F", "G", "H", "I");
+    }
+
+    [Fact]
+    public async Task LoadProjectFromFileAsync_MigrationRequired_ProjectModified()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        var accountingData = new AccountingData { Years = [new AccountingDataYear { Name = 2020 }] };
+        fileSystem.FileExists("the.fileName").Returns(true);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(accountingData.Serialize());
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Completed);
+
+        sut.IsModified.Should().BeTrue();
+    }
+
+    [CulturedFact("en")]
+    public async Task LoadProjectFromFileAsync_UserDoesNotWantSaveCurrentProject_LoadingAborted()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings { SecuredDrives = ["K:\\"] };
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi)
+        {
+            FileName = "old.fileName", IsModified = true
+        };
+        dialogs.ShowMessageBox(
+                Arg.Is<string>(s => s.Contains("Project data has been changed.", StringComparison.Ordinal)),
+                Arg.Any<string>(),
+                MessageBoxButton.YesNo, MessageBoxImage.Question,
+                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
+            .Returns(MessageBoxResult.Cancel);
+        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
+        fileSystem.FileExists("new.fileName").Returns(true);
+
+        (await sut.Awaiting(x => x.LoadFromFileAsync("the.fileName")).Should()
+                .CompleteWithinAsync(10.Seconds()))
+            .Which.Should().Be(OperationResult.Aborted);
+
+        using var _ = new AssertionScope();
+        sut.FileName.Should().Be("old.fileName", "the new file was not loaded");
+        sut.IsModified.Should().BeTrue("changes are (still) not yet saved");
+        fileSystem.DidNotReceive().ReadAllTextFromFile("the.fileName");
+    }
+
     [Fact]
     public async Task LoadProjectFromFileAsync_KnownFileOnSecureDrive_StoreOpenedAndFileLoaded()
     {
@@ -63,18 +278,35 @@ public class ProjectDataTests
     }
 
     [Fact]
-    public void CanDiscardModifiedProject_Cancel_NotSavedAndReturnsFalse()
+    public void TryDiscardModifiedProject_NotModified_ReturnsTrue()
     {
         var windowManager = Substitute.For<IWindowManager>();
         var dialogs = Substitute.For<IDialogs>();
         var fileSystem = Substitute.For<IFileSystem>();
         var processApi = Substitute.For<IProcess>();
+        var sut = new ProjectData(new Settings(), windowManager, dialogs, fileSystem, processApi);
+
+        sut.TryDiscardModifiedProject().Should().BeTrue();
+
+        dialogs.DidNotReceive().ShowMessageBox(
+            Arg.Any<string>(), Arg.Any<string>(),
+            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
+    }
+
+    [Fact]
+    public void TryDiscardModifiedProject_AnswerCancel_NotSavedAndReturnsFalse()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var sut = new ProjectData(new Settings(), windowManager, dialogs, fileSystem, processApi);
         dialogs.ShowMessageBox(
                 Arg.Any<string>(), Arg.Any<string>(),
                 Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
                 Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
             .Returns(MessageBoxResult.Cancel);
-        var sut = new ProjectData(new Settings(), windowManager, dialogs, fileSystem, processApi);
         sut.Load(Samples.SampleProject);
         sut.IsModified = true;
 
@@ -85,6 +317,54 @@ public class ProjectDataTests
             Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
             Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
         fileSystem.DidNotReceive().WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
+    }
+
+    [Fact]
+    public void TryDiscardModifiedProject_AnswerNo_NotSavedAndReturnsTrue()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var sut = new ProjectData(new Settings(), windowManager, dialogs, fileSystem, processApi) { IsModified = true };
+        dialogs.ShowMessageBox(
+                Arg.Any<string>(), Arg.Any<string>(),
+                Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
+            .Returns(MessageBoxResult.No);
+        sut.LoadData(Samples.SampleProject);
+
+        sut.TryDiscardModifiedProject().Should().BeTrue();
+
+        dialogs.Received(1).ShowMessageBox(
+            Arg.Any<string>(), Arg.Any<string>(),
+            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
+        fileSystem.DidNotReceive().WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
+    }
+
+    [Fact]
+    public void TryDiscardModifiedProject_AnswerYes_SavedAndReturnsTrue()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var sut = new ProjectData(new Settings(), windowManager, dialogs, fileSystem, processApi) { IsModified = true };
+        dialogs.ShowMessageBox(
+                Arg.Any<string>(), Arg.Any<string>(),
+                Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
+            .Returns(MessageBoxResult.Yes);
+        sut.LoadData(Samples.SampleProject);
+
+        sut.TryDiscardModifiedProject().Should().BeTrue();
+
+        dialogs.Received(1).ShowMessageBox(
+            Arg.Any<string>(), Arg.Any<string>(),
+            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
+            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
+        fileSystem.Received(1).WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
     }
 
     [Fact]
@@ -122,6 +402,52 @@ public class ProjectDataTests
 
         dialogs.Received(1).ShowSaveFileDialog(Arg.Any<string>());
         fileSystem.DidNotReceive().WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
+    }
+
+    [Fact]
+    public void SaveProject_AutoSaveExisting_AutoSaveFileDeleted()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings();
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        fileSystem.GetLastWriteTime(Arg.Any<string>())
+            .Returns(new DateTime(2020, 2, 29, 18, 45, 56, DateTimeKind.Local));
+        const string fileName = "project.name";
+        fileSystem.FileExists(fileName + "~").Returns(true);
+        sut.LoadData(Samples.SampleProject);
+        sut.FileName = fileName;
+
+        sut.SaveProject();
+
+        fileSystem.DidNotReceive().FileMove(Arg.Any<string>(), Arg.Any<string>());
+        fileSystem.Received(1).WriteAllTextIntoFile(fileName, Arg.Any<string>());
+        fileSystem.Received(1).FileDelete(fileName + "~");
+    }
+
+    [Fact]
+    public void SaveProject_ProjectExisting_SavedAfterBackup()
+    {
+        var windowManager = Substitute.For<IWindowManager>();
+        var dialogs = Substitute.For<IDialogs>();
+        var fileSystem = Substitute.For<IFileSystem>();
+        var processApi = Substitute.For<IProcess>();
+        var settings = new Settings();
+        var sut = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
+        fileSystem.GetLastWriteTime(Arg.Any<string>())
+            .Returns(new DateTime(2020, 2, 29, 18, 45, 56, DateTimeKind.Local));
+        const string fileName = "project.name";
+        fileSystem.FileExists(fileName).Returns(true);
+        sut.LoadData(Samples.SampleProject);
+        sut.FileName = fileName;
+
+        sut.SaveProject();
+
+        fileSystem.Received(1).FileMove(fileName, fileName + ".20200229184556");
+        fileSystem.Received(1).WriteAllTextIntoFile(fileName, Arg.Any<string>());
+        fileSystem.DidNotReceive().FileDelete(Arg.Any<string>());
     }
 
     [Fact]
