+++ tests/SimpleAccounting.UnitTests/Presentation/ShellViewModelTests.cs
@@ -365,267 +365,6 @@ public partial class ShellViewModelTests
     }
 
     [Fact]
-    public void CanDiscardModifiedProject_AnswerNo_NotSavedAndReturnsTrue()
-    {
-        var sut = CreateSut(out var dialogs, out IFileSystem fileSystem);
-        sut.ProjectData.IsModified = true;
-        dialogs.ShowMessageBox(
-                Arg.Any<string>(), Arg.Any<string>(),
-                Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
-            .Returns(MessageBoxResult.No);
-        sut.ProjectData.Load(Samples.SampleProject);
-
-        sut.ProjectData.CanDiscardModifiedProject().Should().BeTrue();
-
-        dialogs.Received(1).ShowMessageBox(
-            Arg.Any<string>(), Arg.Any<string>(),
-            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
-        fileSystem.DidNotReceive().WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
-    }
-
-    [Fact]
-    public void CanDiscardModifiedProject_AnswerYes_SavedAndReturnsTrue()
-    {
-        var sut = CreateSut(out var messageBox, out IFileSystem fileSystem);
-        sut.ProjectData.IsModified = true;
-        messageBox.ShowMessageBox(
-                Arg.Any<string>(), Arg.Any<string>(),
-                Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
-            .Returns(MessageBoxResult.Yes);
-        sut.ProjectData.Load(Samples.SampleProject);
-
-        sut.ProjectData.CanDiscardModifiedProject().Should().BeTrue();
-
-        messageBox.Received(1).ShowMessageBox(
-            Arg.Any<string>(), Arg.Any<string>(),
-            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
-        fileSystem.Received(1).WriteAllTextIntoFile(Arg.Any<string>(), Arg.Any<string>());
-    }
-
-    [Fact]
-    public void CanDiscardModifiedProject_NotModified_ReturnsTrue()
-    {
-        var sut = CreateSut(out IDialogs dialogs);
-
-        sut.ProjectData.CanDiscardModifiedProject().Should().BeTrue();
-
-        dialogs.DidNotReceive().ShowMessageBox(
-            Arg.Any<string>(), Arg.Any<string>(),
-            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
-    }
-
-    [Fact]
-    public async Task LoadProjectFromFileAsync_HappyPath_FileLoaded()
-    {
-        var sut = CreateSut(out IFileSystem fileSystem);
-        fileSystem.FileExists("the.fileName").Returns(true);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        using var _ = new AssertionScope();
-        sut.ProjectData.FileName.Should().Be("the.fileName");
-        sut.ProjectData.IsModified.Should().BeFalse();
-        sut.ProjectData.Settings.RecentProject.Should().Be("the.fileName");
-        sut.ProjectData.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
-        fileSystem.Received(1).ReadAllTextFromFile("the.fileName");
-    }
-
-    [CulturedFact("en")]
-    public async Task LoadProjectFromFileAsync_UserWantsAutoSaveFile_AutoSaveFileLoaded()
-    {
-        var sut = CreateSut(out var dialogs, out IFileSystem fileSystem);
-        dialogs.ShowMessageBox(
-                Arg.Is<string>(s => s.Contains("automatically created backup file", StringComparison.Ordinal)),
-                Arg.Any<string>(),
-                MessageBoxButton.YesNo, MessageBoxImage.Question,
-                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
-            .Returns(MessageBoxResult.Yes);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-        fileSystem.FileExists("the.fileName").Returns(true);
-        fileSystem.FileExists("the.fileName~").Returns(true);
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(5.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        using var _ = new AssertionScope();
-        sut.ProjectData.FileName.Should().Be("the.fileName");
-        sut.ProjectData.IsModified.Should().BeTrue("changes are (still) not yet saved");
-        sut.ProjectData.Settings.RecentProject.Should().Be("the.fileName");
-        sut.ProjectData.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
-        fileSystem.Received(1).ReadAllTextFromFile("the.fileName~");
-    }
-
-    [CulturedFact("en")]
-    public async Task LoadProjectFromFileAsync_UserDoesNotWantAutoSaveFileExists_AutoSaveFileLoaded()
-    {
-        var sut = CreateSut(out var dialogs, out IFileSystem fileSystem);
-        dialogs.ShowMessageBox(
-                Arg.Is<string>(s => s.Contains("automatically created backup file", StringComparison.Ordinal)),
-                Arg.Any<string>(),
-                MessageBoxButton.YesNo, MessageBoxImage.Question,
-                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
-            .Returns(MessageBoxResult.No);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-        fileSystem.FileExists("the.fileName").Returns(true);
-        fileSystem.FileExists("the.fileName~").Returns(true);
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        using var _ = new AssertionScope();
-        sut.ProjectData.FileName.Should().Be("the.fileName");
-        sut.ProjectData.IsModified.Should().BeFalse();
-        sut.ProjectData.Settings.RecentProject.Should().Be("the.fileName");
-        sut.ProjectData.Settings.RecentProjects.OfType<string>().Should().Equal("the.fileName");
-        fileSystem.Received(1).ReadAllTextFromFile("the.fileName");
-        fileSystem.Received(1).FileDelete("the.fileName~");
-    }
-
-    [Theory]
-    [InlineData("Cryptomator File System")]
-    [InlineData("cryptoFs")]
-    public async Task LoadProjectFromFileAsync_NewFileOnSecureDrive_StoreOpenedAndFileLoaded(
-        string cryptoDriveIdentifier)
-    {
-        var sut = CreateSut(out var dialogs, out IFileSystem fileSystem);
-        fileSystem.FileExists(Arg.Is("K:\\the.fileName")).Returns(true);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-        fileSystem.GetDrives().Returns(
-            _ =>
-            {
-                var func1 = new Func<string>(() => "Normal");
-                var func2 = new Func<string>(() => cryptoDriveIdentifier);
-                return new[] { (FilePath: "C:\\", GetFormat: func1), (FilePath: "K:\\", GetFormat: func2) };
-            });
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("K:\\the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        sut.ProjectData.Settings.SecuredDrives.Should().BeEquivalentTo(new object[] { "K:\\" });
-        dialogs.DidNotReceive().ShowMessageBox(
-            Arg.Is<string>(s => s.Contains("Cryptomator", StringComparison.Ordinal)),
-            Arg.Any<string>(),
-            Arg.Any<MessageBoxButton>(), Arg.Any<MessageBoxImage>(),
-            Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>());
-        fileSystem.Received(1).ReadAllTextFromFile("K:\\the.fileName");
-    }
-
-    [Fact]
-    public async Task LoadProjectFromFileAsync_FullRecentList_NewFileOnTop()
-    {
-        var sut = CreateSut(out IFileSystem fileSystem);
-        sut.ProjectData.Settings.RecentProjects =
-        [
-            "A",
-            "B",
-            "C",
-            "D",
-            "E",
-            "F",
-            "G",
-            "H",
-            "I",
-            "J"
-        ];
-        fileSystem.FileExists("the.fileName").Returns(true);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        sut.ProjectData.Settings.RecentProjects.OfType<string>().Should()
-            .Equal("the.fileName", "A", "B", "C", "D", "E", "F", "G", "H", "I");
-    }
-
-    [Fact]
-    public async Task LoadProjectFromFileAsync_MigrationRequired_ProjectModified()
-    {
-        var sut = CreateSut(out IFileSystem fileSystem);
-        var accountingData = new AccountingData { Years = [new AccountingDataYear { Name = 2020 }] };
-        fileSystem.FileExists("the.fileName").Returns(true);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(accountingData.Serialize());
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Completed);
-
-        sut.ProjectData.IsModified.Should().BeTrue();
-    }
-
-    [CulturedFact("en")]
-    public async Task LoadProjectFromFileAsync_UserDoesNotWantSaveCurrentProject_LoadingAborted()
-    {
-        var sut = CreateSut(out var dialogs, out IFileSystem fileSystem);
-        sut.ProjectData.FileName = "old.fileName";
-        sut.ProjectData.IsModified = true;
-        dialogs.ShowMessageBox(
-                Arg.Is<string>(s => s.Contains("Project data has been changed.", StringComparison.Ordinal)),
-                Arg.Any<string>(),
-                MessageBoxButton.YesNo, MessageBoxImage.Question,
-                Arg.Any<MessageBoxResult>(), Arg.Any<MessageBoxOptions>())
-            .Returns(MessageBoxResult.Cancel);
-        fileSystem.ReadAllTextFromFile(Arg.Any<string>()).Returns(new AccountingData().Serialize());
-        fileSystem.FileExists("new.fileName").Returns(true);
-
-        (await sut.Awaiting(x => x.ProjectData.LoadFromFileAsync("the.fileName")).Should()
-                .CompleteWithinAsync(10.Seconds()))
-            .Which.Should().Be(OperationResult.Aborted);
-
-        using var _ = new AssertionScope();
-        sut.ProjectData.FileName.Should().Be("old.fileName", "the new file was not loaded");
-        sut.ProjectData.IsModified.Should().BeTrue("changes are (still) not yet saved");
-        fileSystem.DidNotReceive().ReadAllTextFromFile("the.fileName");
-    }
-
-    [Fact]
-    public void SaveProject_AutoSaveExisting_AutoSaveFileDeleted()
-    {
-        var sut = CreateSut(out IFileSystem fileSystem);
-        fileSystem.GetLastWriteTime(Arg.Any<string>())
-            .Returns(new DateTime(2020, 2, 29, 18, 45, 56, DateTimeKind.Local));
-        var fileName = "project.name";
-        fileSystem.FileExists(fileName + "~").Returns(true);
-        sut.ProjectData.Load(Samples.SampleProject);
-        sut.ProjectData.FileName = fileName;
-
-        sut.ProjectData.SaveProject();
-
-        fileSystem.DidNotReceive().FileMove(Arg.Any<string>(), Arg.Any<string>());
-        fileSystem.Received(1).WriteAllTextIntoFile(fileName, Arg.Any<string>());
-        fileSystem.Received(1).FileDelete(fileName + "~");
-    }
-
-    [Fact]
-    public void SaveProject_ProjectExisting_SavedAfterBackup()
-    {
-        var sut = CreateSut(out IFileSystem fileSystem);
-        fileSystem.GetLastWriteTime(Arg.Any<string>())
-            .Returns(new DateTime(2020, 2, 29, 18, 45, 56, DateTimeKind.Local));
-        var fileName = "project.name";
-        fileSystem.FileExists(fileName).Returns(true);
-        sut.ProjectData.Load(Samples.SampleProject);
-        sut.ProjectData.FileName = fileName;
-
-        sut.ProjectData.SaveProject();
-
-        fileSystem.Received(1).FileMove(fileName, fileName + ".20200229184556");
-        fileSystem.Received(1).WriteAllTextIntoFile(fileName, Arg.Any<string>());
-        fileSystem.DidNotReceive().FileDelete(Arg.Any<string>());
-    }
-
-    [Fact]
     public void ShowInactiveAccounts_SetTrue_InactiveAccountsGetVisible()
     {
         var sut = CreateSut();
@@ -759,28 +498,6 @@ public partial class ShellViewModelTests
         var reportFactory = Substitute.For<IReportFactory>();
         var clock = Substitute.For<IClock>();
         applicationUpdate = Substitute.For<IApplicationUpdate>();
-        dialogs = Substitute.For<IDialogs>();
-        var fileSystem = Substitute.For<IFileSystem>();
-        var processApi = Substitute.For<IProcess>();
-        var settings = new Settings();
-        var projectData = new ProjectData(settings, windowManager, dialogs, fileSystem, processApi);
-        var accountsViewModel = new AccountsViewModel(windowManager, projectData);
-        var sut =
-            new ShellViewModel(
-                projectData, busy,
-                new MenuViewModel(projectData, busy, reportFactory, clock, processApi, dialogs),
-                new FullJournalViewModel(projectData), new AccountJournalViewModel(projectData),
-                accountsViewModel, applicationUpdate);
-        return sut;
-    }
-
-    private static ShellViewModel CreateSut(out IDialogs dialogs)
-    {
-        var busy = Substitute.For<IBusy>();
-        var windowManager = Substitute.For<IWindowManager>();
-        var reportFactory = Substitute.For<IReportFactory>();
-        var clock = Substitute.For<IClock>();
-        var applicationUpdate = Substitute.For<IApplicationUpdate>();
         dialogs = Substitute.For<IDialogs>();
         var fileSystem = Substitute.For<IFileSystem>();
         var processApi = Substitute.For<IProcess>();
